[tf.data] move `StructuredFunctionWrapper` into a common module
refactor using structured_function module

use the new API location of structured_function

move DEBUG_MODE to dataset_ops
